refactor: remove FilePreviewModal and PDF.js dependency
- Removed file preview modal component and its test file
- Updated cloud storage components to open files in new tabs
- Removed file preview related state and actions from useModalStore
- Removed PDF.js dependency from package.json
- Updated layout to remove FilePreviewModal import
