+++ package.json
@@ -74,7 +74,6 @@
     "next": "^15.3.5",
     "next-themes": "^0.4.4",
     "papaparse": "^5.4.1",
-    "pdfjs-dist": "^5.4.54",
     "posthog-js": "^1.266.0",
     "posthog-node": "^5.5.0",
     "react": "^18.3.1",
